Modularized some code, and improved Miku's code
