+++ src/main.py
@@ -3,7 +3,7 @@ import flet as ft
 from main_ui import main_app
 from setup import before_main_app
 
-
+# TODO: Implement a feature to disable/enable debug mode
 DEBUG = True
