+++ src/utilities/debug.py
@@ -1,0 +1,17 @@
+import ctypes
+
+
+def debug_msg(msg: str, handler: str = "DEBUG", debug: bool = False):
+    if debug:
+        print(f"[{handler}] {msg}")
+        
+def get_full_username():
+    GetUserNameEx = ctypes.windll.secur32.GetUserNameExW
+    NameDisplay = 3  # NameDisplay gives full name
+
+    size = ctypes.pointer(ctypes.c_ulong(0))
+    GetUserNameEx(NameDisplay, None, size)
+
+    name_buffer = ctypes.create_unicode_buffer(size.contents.value)
+    GetUserNameEx(NameDisplay, name_buffer, size)
+    return name_buffer.value
